Moved coverage report generation into a separate module
